+++ src/main/java/net/frozenblock/lib/image_transfer/client/ServerTexture.java
@@ -32,6 +32,7 @@ import java.util.concurrent.CompletableFuture;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
+import net.frozenblock.lib.config.frozenlib_config.FrozenLibConfig;
 import net.frozenblock.lib.image_transfer.FileTransferPacket;
 import net.minecraft.Util;
 import net.minecraft.client.Minecraft;
@@ -135,7 +136,9 @@ public class ServerTexture extends SimpleTexture implements Tickable {
             nativeImage = this.load(fileInputStream);
             fileInputStream.close();
         } else {
-			ClientPlayNetworking.send(FileTransferPacket.createRequest(this.destPath, this.fileName));
+			if (FrozenLibConfig.FILE_TRANSFER_CLIENT) {
+				ClientPlayNetworking.send(FileTransferPacket.createRequest(this.destPath, this.fileName));
+			}
             nativeImage = null;
         }